trellis-worker/requirements.txt
# Bjorq 3D Worker — Python dependencies
# PyTorch is installed separately with CUDA index URL by the installer
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
huggingface-hub>=0.26.0
Pillow>=10.0.0
